# [BE] refactor: ImageKitProperties 위치 변경
